requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
aiortc==1.9.0
annotated-types==0.7.0
bcrypt==4.1.3
cassandra-driver==3.29.2
celery==5.4.0
dataclasses==0.6
dataclasses-json==0.6.7
fastapi==0.113.0
gTTS==2.5.3
gradio_client==1.3.0
langchain
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-google-genai==2.0.0
langchain-text-splitters==0.3.0
langsmith==0.1.120
loguru==0.7.2
multidict==6.0.5
ollama==0.3.2
openai==1.43.1
pandas==2.2.2
PyAudio==0.2.14
pydantic==2.9.0
pydub==0.25.1
PyJWT==2.8.0
pypdf==4.3.1
python-dotenv==1.0.1
python-multipart==0.0.9
pillow==10.4.0
redis==5.0.8
requests==2.32.3
SpeechRecognition==3.10.4
starlette==0.38.4
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
typing_extensions==4.12.2
ulid==1.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.19.0
websockets==12.0

